refactor(cli): move watch module under commands
