+++ crates/tsz-cli/src/commands/mod.rs
@@ -1,2 +1,3 @@
 pub mod build;
 pub mod help;
+pub mod watch;
