+++ crates/tsz-cli/src/lib.rs
@@ -14,9 +14,9 @@ pub mod project;
 pub mod project_refs;
 pub mod reporting;
 pub mod tracing_config;
-pub mod watch;
 pub use commands::build;
 pub use commands::help;
+pub use commands::watch;
 pub use project::fs;
 pub use reporting::reporter;
 pub use reporting::trace;
